% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simitation.R
\name{buildsim.binary}
\alias{buildsim.binary}
\title{Internal function for Simulation for Binary Data}
\usage{
buildsim.binary(the.formula, the.variable, n, num.experiments = 1)
}
\arguments{
\item{the.formula}{A character string specifying the distribution function, e.g., "binary(0.5)".}

\item{the.variable}{A character string naming the variable in the generated data.}

\item{n}{An integer specifying the number of data points to generate for each experiment.}

\item{num.experiments}{An integer specifying the number of experiments to simulate. Default is 1.}
}
\value{
A data frame with simulated binary values based on the given formula.
}
\description{
This function is designed to generate binary data based on the provided formula.
It is an internal function and is not meant for end-users.
}
