% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simitation.R
\name{buildsim.binomial}
\alias{buildsim.binomial}
\title{Internal function for Simulation for Binomial Data}
\usage{
buildsim.binomial(the.formula, the.variable, n, num.experiments = 1)
}
\arguments{
\item{the.formula}{A character string specifying the distribution function, e.g., "Bin(10, 0.5)".}

\item{the.variable}{A character string naming the variable in the generated data.}

\item{n}{An integer specifying the number of data points to generate for each experiment.}

\item{num.experiments}{An integer specifying the number of experiments to simulate. Default is 1.}
}
\value{
A data frame with simulated binomial values based on the given formula.
}
\description{
This internal function is designed to generate binomial data based on the
provided formula. It is not intended for direct usage by end-users.
}
