% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simitation.R
\name{buildsim.normal}
\alias{buildsim.normal}
\title{Internal Simulation for Normally Distributed Data}
\usage{
buildsim.normal(the.formula, the.variable, n, num.experiments)
}
\arguments{
\item{the.formula}{A character string specifying the normal distribution function, e.g., "N(0,1)".}

\item{the.variable}{A character string naming the variable in the generated data.}

\item{n}{An integer specifying the number of data points to generate for each experiment.}

\item{num.experiments}{An integer specifying the number of experiments to simulate.}
}
\value{
A data frame with simulated normally distributed values based on the given formula.
}
\description{
This internal function generates data based on a normal distribution specified
by the provided formula. It is not intended for direct usage by end-users.
}
