% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simitation.R
\name{buildsim.poisson}
\alias{buildsim.poisson}
\title{Internal function for Simulation for Poisson Distributed Data}
\usage{
buildsim.poisson(the.formula, the.variable, n, num.experiments)
}
\arguments{
\item{the.formula}{A character string specifying the Poisson distribution function, e.g., "poisson(3)".}

\item{the.variable}{A character string naming the variable in the generated data.}

\item{n}{An integer specifying the number of data points to generate for each experiment.}

\item{num.experiments}{An integer specifying the number of experiments to simulate.}
}
\value{
A data frame with simulated Poisson distributed values based on the given formula.
}
\description{
This internal function generates data based on a Poisson distribution specified
by the provided formula. It is not intended for direct usage by end-users.
}
