% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simitation.R
\name{buildsim.sample}
\alias{buildsim.sample}
\title{Internal function for Sampling Function}
\usage{
buildsim.sample(
  the.formula,
  the.variable,
  n,
  num.experiments,
  value.split = ",",
  symbol.open.paren = "(",
  symbol.close.paren = ")"
)
}
\arguments{
\item{the.formula}{A character string specifying the sampling formula, e.g., "sample(('Red', 'Green', 'Blue'), (0.5, 0.3, 0.2))".}

\item{the.variable}{A character string naming the variable in the generated data.}

\item{n}{An integer specifying the number of data points to generate for each experiment.}

\item{num.experiments}{An integer specifying the number of experiments to simulate.}

\item{value.split}{A character used to split the values in the sample.}

\item{symbol.open.paren}{A character specifying the opening parenthesis.}

\item{symbol.close.paren}{A character specifying the closing parenthesis.}
}
\value{
A data frame with sampled values based on the given formula.
}
\description{
This internal function generates samples based on the specified distributions and probabilities.
It is not intended for direct usage by end-users.
}
