% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simitation.R
\name{buildsim.uniform}
\alias{buildsim.uniform}
\title{Internal function for Simulation for Uniform Distributed Data}
\usage{
buildsim.uniform(the.formula, the.variable, n, num.experiments = 1)
}
\arguments{
\item{the.formula}{A character string specifying the Uniform distribution function, e.g., "U(0,1)".}

\item{the.variable}{A character string naming the variable in the generated data.}

\item{n}{An integer specifying the number of data points to generate for each experiment.}

\item{num.experiments}{An integer specifying the number of experiments to simulate.}
}
\value{
A data frame with simulated Uniform distributed values based on the given formula.
}
\description{
This internal function generates data based on a Uniform distribution specified
by the provided formula. It is not intended for direct usage by end-users.
}
