% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simitation.R
\name{identify.distribution}
\alias{identify.distribution}
\title{Internal function for Distribution Identification}
\usage{
\method{identify}{distribution}(
  dat = NULL,
  the.step,
  n,
  num.experiments,
  step.split = "~",
  value.split = ","
)
}
\arguments{
\item{dat}{Optional data table for generating data.}

\item{the.step}{A character string specifying the formula for simulation.}

\item{n}{An integer specifying the number of data points to generate.}

\item{num.experiments}{An integer specifying the number of experiments to simulate.}

\item{step.split}{A character indicating the delimiter for splitting the step formula.}

\item{value.split}{A character used to split values in certain distributions.}
}
\value{
A data table with simulated values based on the identified distribution.
}
\description{
This internal function identifies the type of distribution based on a given formula and simulates
data accordingly. It is not intended for direct usage by end-users.
}
