% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simitation.R
\name{internal.chisq.test.gf}
\alias{internal.chisq.test.gf}
\title{Internal function for Chi-Squared Test of Goodness of Fit}
\usage{
internal.chisq.test.gf(x, hypothesized.probs = NULL, correct = TRUE)
}
\arguments{
\item{x}{A categorical variable.}

\item{hypothesized.probs}{Hypothesized probabilities for each level of x.}

\item{correct}{A logical indicating if a continuity correction should be applied.}
}
\value{
A data frame containing the test statistic, degrees of freedom, and p-value.
}
\description{
Computes the chi-squared test for the given categorical data.
}
