% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simitation.R
\name{internal.chisq.test.ind}
\alias{internal.chisq.test.ind}
\title{Internal function for Chi-Squared Test of Independence}
\usage{
internal.chisq.test.ind(the.data, group.name, value.name, correct = TRUE)
}
\arguments{
\item{the.data}{The data table.}

\item{group.name}{Group variable name.}

\item{value.name}{Value variable name.}

\item{correct}{A logical indicating if a continuity correction should be applied.}
}
\value{
A data frame containing the test statistic, degrees of freedom, and p-value.
}
\description{
Computes the chi-squared test for the given data.
}
