% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simitation.R
\name{internal.prop.test}
\alias{internal.prop.test}
\title{Internal function for One-sample Proportions Test}
\usage{
internal.prop.test(
  x,
  p = NULL,
  alternative = c("two.sided", "less", "greater"),
  conf.level = 0.95,
  correct = TRUE
)
}
\arguments{
\item{x}{A binary variable.}

\item{...}{Other parameters passed to prop.test function.}
}
\value{
A data frame with test results.
}
\description{
Computes the test for proportion for the given binary data.
}
