% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simitation.R
\name{internal.prop2.test}
\alias{internal.prop2.test}
\title{Internal function for Two-sample Proportions Test}
\usage{
internal.prop2.test(
  x,
  y,
  p = NULL,
  alternative = c("two.sided", "less", "greater"),
  conf.level = 0.95,
  correct = TRUE,
  na.rm = T
)
}
\arguments{
\item{x}{First binary variable.}

\item{y}{Second binary variable.}

\item{...}{Other parameters passed to prop.test function.}
}
\value{
A data frame with test results.
}
\description{
Computes the test for proportions for two given binary variables.
}
