% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simitation.R
\name{internal.quantiles.mean.sd}
\alias{internal.quantiles.mean.sd}
\title{Internal function for Quantile, Mean, and Standard Deviation Calculation}
\usage{
internal.quantiles.mean.sd(x, the.quantiles, na.rm = T)
}
\arguments{
\item{x}{A numeric vector.}

\item{the.quantiles}{A numeric vector of quantile values.}

\item{na.rm}{A logical indicating if missing values should be removed.}
}
\value{
A data table with summary statistics.
}
\description{
Computes the specified quantiles, mean, and standard deviation for the given data.
}
