% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simitation.R
\name{internal.statistics.one.lm}
\alias{internal.statistics.one.lm}
\title{Internal function for Summary Statistics of Linear Model}
\usage{
internal.statistics.one.lm(the.data, the.formula)
}
\arguments{
\item{the.data}{The data table.}

\item{the.formula}{A formula specifying the linear model.}
}
\value{
A list containing the following components:

\itemize{
  \item{\code{coef.table}}{A data table containing the coefficients, their estimates, standard error, t-values, and associated p-values.}
  \item{\code{summary.stats}}{A data table containing summary statistics for the linear model. This includes the sigma, degrees of freedom, root mean square error (rse), R-squared, adjusted R-squared, F-statistic, F-statistic degrees of freedom, and the p-value for the F-statistic.}
}
}
\description{
Computes the summary statistics for the linear model fit on the given data.
}
