% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simitation.R
\name{internal.statistics.one.logistic}
\alias{internal.statistics.one.logistic}
\title{Internal function for Summary Statistics of Logistic Regression}
\usage{
internal.statistics.one.logistic(the.data, the.formula)
}
\arguments{
\item{the.data}{The data table.}

\item{the.formula}{A formula specifying the logistic regression model.}
}
\value{
A list containing the coefficient table and summary statistics.
}
\description{
Computes the summary statistics for the logistic regression fit on the given data.
}
