% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simitation.R
\name{internal.statistics.onelm}
\alias{internal.statistics.onelm}
\title{Internal function for Summary Statistics of Linear Model}
\usage{
internal.statistics.onelm(the.data, the.formula)
}
\arguments{
\item{the.data}{The data table.}

\item{the.formula}{A formula specifying the linear model.}
}
\value{
A list containing the coefficient table and summary statistics.
}
\description{
Computes the summary statistics for the linear model fit on the given data.
}
