% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simitation.R
\name{quantile.dt}
\alias{quantile.dt}
\title{Internal function for Quantiles Calculation}
\usage{
\method{quantile}{dt}(x, probs, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector.}

\item{probs}{A numeric vector of quantile values.}

\item{na.rm}{A logical indicating if missing values should be removed.}
}
\value{
A data table with quantile values.
}
\description{
Computes the specified quantiles for the given data.
}
