% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simitation.R
\name{sim.chisq.gf}
\alias{sim.chisq.gf}
\title{sim.chisq.gf}
\usage{
sim.chisq.gf(
  n,
  values,
  prob = NULL,
  num.experiments = 1,
  experiment.name = "experiment",
  value.name = "x",
  seed = 91,
  vstr = 3.6
)
}
\arguments{
\item{n:}{A numeric value for the number of observations in each experiment.}

\item{values:}{A vector of values specifying the sample space.}

\item{prob:}{A vector of probabilities used to simulate the values.}

\item{num.experiments:}{A numeric value representing the number of simulated experiments.}

\item{experiment.name:}{A character value providing the name for the column identifying the experiment.}

\item{value.name:}{A character value providing the name for the simulated values.}

\item{seed:}{A single numeric value, interpreted as an integer, or NULL.   See help(set.seed).}

\item{vstr:}{A character string containing a version number, e.g., "1.6.2". The default RNG configuration of the current R version is used if vstr is greater than the current version.  See help(set.seed).}
}
\description{
sim.chisq.gf
}
