% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simitation.R
\name{sim.statistics.logistic}
\alias{sim.statistics.logistic}
\title{sim.statistics.lm}
\usage{
sim.statistics.logistic(simdat, the.formula, grouping.variables)
}
\arguments{
\item{simdat}{Data for use in multivariable regression models across one or more experiments.  Structure is in the form returned by the function simitation::simulation.steps().}

\item{the.formula}{A formula object or character value specifying the formula for the regression model.}

\item{grouping.variables}{A character vector of column names from simdat on which to group the data.  The intended regression model will be fit in groups based on this selection.}
}
\description{
sim.statistics.lm
}
