% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simitation.R
\name{analyze.simstudy.t}
\alias{analyze.simstudy.t}
\title{analyze.simstudy.t}
\usage{
analyze.simstudy.t(
  test.statistics.t,
  alternative = c("two.sided", "less", "greater"),
  conf.level = 0.95,
  the.quantiles = c(0.025, 0.1, 0.25, 0.5, 0.75, 0.9, 0.975)
)
}
\arguments{
\item{test.statistics.t:}{Summary information for fitting one-sample t tests.  Structure is in the form returned by the function simitation::sim.t.test().}

\item{alternative:}{See help(t.test).}

\item{conf.level:}{See help(t.test)}

\item{the.quantiles:}{A numeric vector of values between 0 and 1.  Summary statistics to analyze the tests will return the specified quantiles.}
}
\description{
analyze.simstudy.t
}
