% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simitation.R
\name{internal.statistics.one.lm}
\alias{internal.statistics.one.lm}
\title{Internal function for Summary Statistics of Linear Model}
\usage{
internal.statistics.one.lm(the.data, the.formula)
}
\arguments{
\item{the.data}{The data table.}

\item{the.formula}{A formula specifying the linear model.}
}
\description{
Computes the summary statistics for the linear model fit on the given data.
}
