% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simitation.R
\name{internal.t.test}
\alias{internal.t.test}
\title{Internal function for One-sample t-test}
\usage{
internal.t.test(
  x,
  alternative = c("two.sided", "less", "greater"),
  mu = 0,
  paired = FALSE,
  var.equal = FALSE,
  conf.level = 0.95
)
}
\arguments{
\item{x}{A numeric vector.}

\item{...}{Other parameters passed to t.test function.}
}
\value{
A data frame with test results.
}
\description{
Computes the one-sample t-test for the given data.
}
