% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simitation.R
\name{internal.t2.test}
\alias{internal.t2.test}
\title{Internal function for Two-sample t-test}
\usage{
internal.t2.test(
  x,
  y,
  alternative = c("two.sided", "less", "greater"),
  mu = 0,
  paired = FALSE,
  var.equal = FALSE,
  conf.level = 0.95
)
}
\arguments{
\item{x}{First numeric vector.}

\item{y}{Second numeric vector.}

\item{...}{Other parameters passed to t.test function.}
}
\value{
A data frame with test results.
}
\description{
Computes the two-sample t-test for the given data.
}
