% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simitation.R
\name{sim.chisq.test.ind}
\alias{sim.chisq.test.ind}
\title{sim.chisq.test.ind}
\usage{
sim.chisq.test.ind(
  simdat.chisq.ind,
  correct = T,
  experiment.name = "experiment",
  group.name = "variable",
  value.name = "value"
)
}
\arguments{
\item{simdat.chisq.ind}{Data for use in chi squared tests of independence across one or more experiments.  Structure is in the form returned by the function simitation::sim.chisq.ind().}

\item{correct}{See help(chisq.test).}

\item{experiment.name}{A character value providing the name of the column identifying the experiment.}

\item{group.name}{A character value providing the name of the column of the group labels.}

\item{value.name}{A character value providing the name of the column identifying the values.}
}
\description{
sim.chisq.test.ind
}
