% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simitation.R
\name{sim.prop.test}
\alias{sim.prop.test}
\title{sim.prop.test}
\usage{
sim.prop.test(
  simdat.prop,
  p = NULL,
  alternative = c("two.sided", "less", "greater"),
  conf.level = 0.95,
  correct = T,
  experiment.name = "experiment",
  value.name = "x"
)
}
\arguments{
\item{simdat.prop}{Data for use in one-sample proportions tests across one or more experiments.  Structure is in the form returned by the function simitation::sim.prop().}

\item{p}{See help(prop.test).}

\item{alternative}{See help(prop.test).}

\item{conf.level}{See help(prop.test).}

\item{correct}{See help(prop.test).}

\item{experiment.name}{A character value providing the name of the column identifying the experiment.}

\item{value.name}{A character value providing the name of the column identifying the values.}
}
\description{
sim.prop.test
}
